final updates & tests
I added explanations to the pybids and nilearn-GLM notebooks. The local build of the image worked in general. However, the removal of the data from sub-04/05/06/08/09/10 doesn't seem to work as it was still in the image. Therefore, the BIDS dataset analyses part of the Nilearn-GLM notebook will crash until that data is deleted. We should check that accordingly.